refactor: sync logic
